Extract catalog schema and types into @tomomai/catalog package
Move the catalog-side contract out of apps/main into a shared workspace
package so the upcoming data service can reuse it:

- songs/tourEvents/tourEventSteps tables + region/difficulty/level/chart
  type enums -> packages/catalog/src/schema.ts
- catalog enum value arrays -> enums.ts
- Region/Difficulty/Level/SongType/MinimalSong/SongBase/SongExtended/
  NoteCounts types -> types.ts
- version metadata (Versions table + helpers) -> metadata.ts
- normalizeName/normalizeGenre/renderLevelPrecise -> name-utils.ts

The original apps/main modules re-export from @tomomai/catalog, so all
existing import sites and drizzle-kit are unaffected. Verified: drizzle
generate reports no schema changes, migrations apply, typecheck/lint/
build pass.
