+++ apps/main/src/lib/db/types.ts
@@ -1,16 +1,7 @@
+// Catalog enums live in @tomomai/catalog (shared with the data service)
+export { REGION_ENUM, DIFFICULTY_ENUM, LEVEL_ENUM, CHART_TYPE_ENUM } from "@tomomai/catalog/enums";
+
 export const LANGUAGE_ENUM = ["en", "en-GB", "ja", "zh-TW", "zh-HK", "zh-CN", "zh-SG", "ko"] as const;
-
-export const REGION_ENUM = ["intl", "jp", "cn"] as const;
-
-export const DIFFICULTY_ENUM = ["basic", "advanced", "expert", "master", "remaster", "utage"] as const;
-
-export const LEVEL_ENUM = [
-  "1", "1+", "2", "2+", "3", "3+", "4", "4+", "5", "5+", "6", "6+",
-  "7", "7+", "8", "8+", "9", "9+", "10", "10+", "11", "11+", "12", "12+",
-  "13", "13+", "14", "14+", "15", "15+", "16", "16+"
-] as const;
-
-export const CHART_TYPE_ENUM = ["std", "dx"] as const;
 
 export const FC_ENUM = ["none", "fc", "fc+", "ap", "ap+"] as const;
