+++ packages/catalog/src/artifact.ts
@@ -1,0 +1,5 @@
+// Zod schemas describing the published catalog artifact and its manifest.
+// Populated when the data service (apps/data) starts publishing; kept here so
+// publisher and consumers validate against the same contract.
+
+export const CATALOG_SCHEMA_VERSION = 1;
