+++ packages/catalog/src/enums.ts
@@ -1,0 +1,11 @@
+export const REGION_ENUM = ["intl", "jp", "cn"] as const;
+
+export const DIFFICULTY_ENUM = ["basic", "advanced", "expert", "master", "remaster", "utage"] as const;
+
+export const LEVEL_ENUM = [
+  "1", "1+", "2", "2+", "3", "3+", "4", "4+", "5", "5+", "6", "6+",
+  "7", "7+", "8", "8+", "9", "9+", "10", "10+", "11", "11+", "12", "12+",
+  "13", "13+", "14", "14+", "15", "15+", "16", "16+"
+] as const;
+
+export const CHART_TYPE_ENUM = ["std", "dx"] as const;
